+++ includes/admin/class-cocart-wc-admin-system-status.php
@@ -4,11 +4,11 @@
  *
  * Adds additional related information to the WooCommerce System Status.
  *
- * @since    2.1.0
- * @version  2.1.2
  * @author   Sébastien Dumont
  * @category Admin
  * @package  CoCart/Admin/System Status
+ * @since    2.1.0
+ * @version  2.1.2
  * @license  GPL-2.0+
  */