+++ includes/admin/views/html-notice-requirement-wp.php
@@ -2,11 +2,11 @@
 /**
  * Admin View: WordPress Requirement Notice.
  *
- * @since    1.2.0
- * @version  2.0.0
  * @author   Sébastien Dumont
  * @category Admin
  * @package  CoCart/Admin/Views
+ * @since    1.2.0
+ * @version  2.3.0
  * @license  GPL-2.0+
  */
 
@@ -16,5 +16,5 @@ if ( ! defined( 'ABSPATH' ) ) {
 }
 ?>
 <div class="notice notice-error">
-	<p><?php echo sprintf( __( 'Sorry, %1$s%3$s%2$s requires WordPress %4$s or higher. Please upgrade your WordPress setup.', 'cart-rest-api-for-woocommerce' ), '<strong>', '</strong>', 'CoCart', COCART_WP_VERSION_REQUIRE ); ?></p>
+	<p><?php echo sprintf( __( 'Sorry, %1$s%3$s%2$s requires WordPress %4$s or higher. Please upgrade your WordPress setup.', 'cart-rest-api-for-woocommerce' ), '<strong>', '</strong>', 'CoCart', CoCart::$required_wp ); ?></p>
 </div>