+++ includes/class-cocart-autoloader.php
@@ -2,10 +2,10 @@
 /**
  * CoCart - Autoloader.
  *
- * @since    2.0.0
  * @author   Sébastien Dumont
  * @category Classes
  * @package  CoCart/Classes/AutoLoader
+ * @since    2.0.0
  * @license  GPL-2.0+
  */