+++ includes/class-cocart-logger.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * CoCart REST API logger
+ * CoCart REST API logger.
  *
  * Handles logging errors.
  *
@@ -9,6 +9,7 @@
  * @package  CoCart/Logger
  * @since    2.1.0
  * @version  2.2.0
+ * @license  GPL-2.0+
  */
 
 if ( ! defined( 'ABSPATH' ) ) {