+++ includes/class-cocart-session-handler.php
@@ -10,7 +10,8 @@
  * @category API
  * @package  CoCart/Session
  * @since    2.1.0
- * @version  2.1.7
+ * @version  2.3.0
+ * @license  GPL-2.0+
  */
 
 if ( ! defined( 'ABSPATH' ) ) {
@@ -72,7 +73,7 @@ class CoCart_Session_Handler extends WC_Session {
 	 *
 	 * @access  public
 	 * @since   2.1.0
-	 * @version 2.1.5
+	 * @version 2.3.0
 	 */
 	public function init() {
 		// Current user ID. If user is NOT logged in then the customer is a guest.
@@ -88,9 +89,9 @@ class CoCart_Session_Handler extends WC_Session {
 		 * When a user is logged out, ensure they have a unique nonce by using the customer/cart ID.
 		 *
 		 * @since   2.1.2
-		 * @version 2.1.5
+		 * @version 2.3.0
 		 */
-		if ( CoCart::is_rest_api_request() && is_numeric( $current_user_id ) && $current_user_id < 1 ) {
+		if ( CoCart_Helpers::is_rest_api_request() && is_numeric( $current_user_id ) && $current_user_id < 1 ) {
 			add_filter( 'nonce_user_logged_out', array( $this, 'nonce_user_logged_out' ) );
 		}
 	} // END init()